+++ R/convert_to_na.R
@@ -48,7 +48,10 @@ convert_to_na <- function(x, ...) {
 convert_to_na.default <- function(x, verbose = TRUE, ...) {
   if (isTRUE(verbose)) {
     message(insight::format_message(
-      sprintf("Converting values into missing values (`NA`) currently not possible for variables of class `%s`.", class(x)[1])
+      sprintf(
+        "Converting values into missing values (`NA`) currently not possible for variables of class `%s`.",
+        class(x)[1]
+      )
     ))
   }
   x
@@ -60,7 +63,7 @@ convert_to_na.default <- function(x, verbose = TRUE, ...) {
 convert_to_na.numeric <- function(x, na = NULL, verbose = TRUE, ...) {
   # if we have a list, use first valid element
   if (is.list(na)) {
-    na <- unlist(na[sapply(na, is.numeric)])
+    na <- unlist(na[vapply(na, is.numeric, FUN.VALUE = TRUE)])
   }
 
   if (is_empty_object(na) || !is.numeric(na)) {
@@ -86,7 +89,7 @@ convert_to_na.numeric <- function(x, na = NULL, verbose = TRUE, ...) {
 convert_to_na.factor <- function(x, na = NULL, drop_levels = FALSE, verbose = TRUE, ...) {
   # if we have a list, use first valid element
   if (is.list(na)) {
-    na <- unlist(na[sapply(na, is.character)])
+    na <- unlist(na[vapply(na, is.character, FUN.VALUE = TRUE)])
   }
 
   if (is_empty_object(na) || (!is.factor(na) && !is.character(na))) {
@@ -142,7 +145,7 @@ convert_to_na.Date <- function(x, na = NULL, verbose = TRUE, ...) {
 convert_to_na.logical <- function(x, na = NULL, verbose = TRUE, ...) {
   # if we have a list, use first valid element
   if (is.list(na)) {
-    na <- unlist(na[sapply(na, is.logical)])
+    na <- unlist(na[vapply(na, is.logical, FUN.VALUE = TRUE)])
   }
 
   if (is_empty_object(na) || !is.logical(na)) {
@@ -180,6 +183,14 @@ convert_to_na.data.frame <- function(x,
     verbose = verbose
   )
 
-  x[select] <- lapply(x[select], convert_to_na, na = na, drop_levels = drop_levels, verbose = verbose, ...)
+  x[select] <- lapply(
+    x[select],
+    convert_to_na,
+    na = na,
+    drop_levels = drop_levels,
+    verbose = verbose,
+    ...
+  )
+
   x
 }