+++ R/data_reshape.R
@@ -95,7 +95,7 @@ data_to_long <- function(data,
 
   if (inherits(data, "tbl_df")) {
     tbl_input <- TRUE
-    data <- as.data.frame(data)
+    data <- as.data.frame(data, stringsAsFactors = FALSE)
   } else {
     tbl_input <- FALSE
   }
@@ -180,22 +180,23 @@ data_to_long <- function(data,
   # if several variable in names_to, split the names either with names_sep
   # or with names_pattern
   if (length(names_to) > 1) {
-    for (i in seq_along(names_to)) {
-      if (is.null(names_pattern)) {
+    if (is.null(names_pattern)) {
+      for (i in seq_along(names_to)) {
         new_vals <- unlist(lapply(
           strsplit(unique(long[[names_to_2]]), names_sep, fixed = TRUE),
           function(x) x[i]
         ))
         long[[names_to[i]]] <- new_vals
-      } else {
-        tmp <- regmatches(
-          unique(long[[names_to_2]]),
-          regexec(names_pattern, unique(long[[names_to_2]]))
-        )
-        tmp <- as.data.frame(do.call(rbind, tmp))[, c(1, i + 1)]
-        names(tmp) <- c(names_to_2, names_to[i])
-        long <- data_join(long, tmp)
       }
+    } else {
+      tmp <- regmatches(
+        unique(long[[names_to_2]]),
+        regexec(names_pattern, unique(long[[names_to_2]]))
+      )
+      tmp <- as.data.frame(do.call(rbind, tmp), stringsAsFactors = FALSE)
+      names(tmp) <- c(names_to_2, names_to)
+      # faster than merge
+      long <- cbind(long, tmp[match(long[[names_to_2]], tmp[[names_to_2]]), -1])
     }
     long[[names_to_2]] <- NULL
   }
@@ -317,7 +318,6 @@ data_to_long <- function(data,
 #'
 #' @inherit data_rename seealso
 #' @export
-
 data_to_wide <- function(data,
                          id_cols = NULL,
                          values_from = "Value",
@@ -354,7 +354,7 @@ data_to_wide <- function(data,
   # Preserve attributes
   if (inherits(data, "tbl_df")) {
     tbl_input <- TRUE
-    data <- as.data.frame(data)
+    data <- as.data.frame(data, stringsAsFactors = FALSE)
   } else {
     tbl_input <- FALSE
   }
@@ -363,12 +363,9 @@ data_to_wide <- function(data,
 
   # Create an id for stats::reshape
   if (is.null(id_cols)) {
-    data[["_Rows"]] <- apply(
-      data[, !names(data) %in% c(values_from, names_from), drop = FALSE],
-      MARGIN = 1,
-      paste,
-      collapse = "_"
-    )
+    row_index <- do.call(paste, c(data[, !names(data) %in% c(values_from, names_from), drop = FALSE], sep = "_"))
+    if (length(row_index) == 0) row_index <- rep("", nrow(data))
+    data[["_Rows"]] <- row_index
     id_cols <- "_Rows"
   }
 
@@ -378,7 +375,7 @@ data_to_wide <- function(data,
   current_colnames <- colnames(data)
   current_colnames <- current_colnames[current_colnames != "_Rows"]
   if (is.null(names_glue)) {
-    future_colnames <- unique(apply(data, 1, function(x) paste(x[c(names_from)], collapse = names_sep)))
+    future_colnames <- unique(do.call(paste, c(data[, names_from, drop = FALSE], sep = names_sep)))
   } else {
     vars <- regmatches(names_glue, gregexpr("\\{\\K[^{}]+(?=\\})", names_glue, perl = TRUE))[[1]]
     tmp_data <- unique(data[, vars])
@@ -409,7 +406,7 @@ data_to_wide <- function(data,
   # stats::reshape works strangely when several variables are in idvar/timevar
   # so we unite all ids in a single temporary column that will be used by
   # stats::reshape
-  data$new_time <- apply(data, 1, function(x) paste(x[names_from], collapse = "_"))
+  data$new_time <- do.call(paste, c(data[, names_from, drop = FALSE], sep = "_"))
   data[, names_from] <- NULL
 
   wide <- stats::reshape(