+++ tests/testthat/test-data_restoretype.R
@@ -2,7 +2,8 @@ test_that("data_restoretype works as expected", {
   data <- data.frame(
     Sepal.Length = c("1", "3", "2"),
     Species = c("setosa", "versicolor", "setosa"),
-    New = c("1", "3", "4")
+    New = c("1", "3", "4"),
+    stringsAsFactors = FALSE
   )
 
   fixed <- data_restoretype(data, reference = iris)