+++ tests/testthat/test-select_nse.R
@@ -1,11 +1,30 @@
-foo <- function(data, select = NULL, exclude = NULL) {
-  .select_nse(select, data, exclude = exclude, ignore_case = FALSE)
+foo <- function(data, select = NULL, exclude = NULL, regex = FALSE) {
+  .select_nse(select, data, exclude = exclude, regex = regex, ignore_case = FALSE)
 }
+
+test_that(".select_nse needs data", {
+  expect_error(foo(select = "Sepal.Length", data = NULL), regexp = "must be provided")
+})
+
+test_that(".select_nse needs a data frame or something coercible to a dataframe", {
+  expect_equal(
+    foo(select = "Sepal.Length", data = as.matrix(head(iris))),
+    "Sepal.Length"
+  )
+  expect_error(
+    foo(select = "Sepal.Length", data = list(1:3, 1:2)),
+    regexp = "must be a data frame"
+  )
+})
 
 test_that(".select_nse: arg 'select' works", {
   expect_equal(
     foo(iris, select = NULL),
     names(iris)
+  )
+  expect_equal(
+    foo(iris, Petal.Length),
+    "Petal.Length"
   )
   expect_equal(
     foo(iris, c("Petal.Length", "Sepal.Width")),
@@ -33,6 +52,10 @@ test_that(".select_nse: arg 'select' works", {
   )
   expect_equal(
     foo(iris, regex("Length$")),
+    c("Sepal.Length", "Petal.Length")
+  )
+  expect_equal(
+    foo(iris, "Len", regex = TRUE),
     c("Sepal.Length", "Petal.Length")
   )
 })